search: score policy on the boolean node classes
