+++ libs/iresearch/include/iresearch/search/common/boolean_groups.hpp
@@ -59,6 +59,13 @@ class OrGroup {
     });
   }
 
+  doc_id_t Fill(doc_id_t min, doc_id_t max, uint64_t* IRS_RESTRICT words,
+                score_t* IRS_RESTRICT scores) {
+    return _leaves.Visit(max, [&](auto& leaf) IRS_FORCE_INLINE {
+      return leaf.Fill(min, max, words, scores);
+    });
+  }
+
  private:
   Leaves _leaves;
 };
