+++ libs/iresearch/include/iresearch/search/common/conjunction_of.hpp
@@ -49,15 +49,18 @@ Result BuildConjunctionOf(std::span<const Term> terms,
   return BuildConjunction<Result, Term>(
     terms, filters, field, segment, interrogations,
     []<typename Lead, typename Others>(auto&& lead, auto&& others) -> Result {
-      using Node = lead::BooleanSparse<Lead, Others, utils::Empty>;
+      using Node =
+        lead::BooleanSparse<Lead, Others, utils::Empty, utils::Empty>;
       if constexpr (kFilled) {
         return memory::make_managed<fill::ByWalkDocs<Node>>(
           std::piecewise_construct, std::forward<decltype(lead)>(lead),
-          std::forward<decltype(others)>(others), std::forward_as_tuple());
+          std::forward<decltype(others)>(others), std::forward_as_tuple(),
+          std::forward_as_tuple());
       } else {
         return memory::make_managed<lead::Impl<Node>>(
           std::piecewise_construct, std::forward<decltype(lead)>(lead),
-          std::forward<decltype(others)>(others), std::forward_as_tuple());
+          std::forward<decltype(others)>(others), std::forward_as_tuple(),
+          std::forward_as_tuple());
       }
     });
 }
