--- libs/iresearch/include/iresearch/search/fill/make_window_disjunction_scored.cpp
@@ -1,29 +1,0 @@
-////////////////////////////////////////////////////////////////////////////////
-/// DISCLAIMER
-///
-/// Copyright 2026 SereneDB GmbH, Berlin, Germany
-///
-/// Licensed under the Apache License, Version 2.0 (the "License");
-/// you may not use this file except in compliance with the License.
-/// You may obtain a copy of the License at
-///
-///     http://www.apache.org/licenses/LICENSE-2.0
-///
-/// Unless required by applicable law or agreed to in writing, software
-/// distributed under the License is distributed on an "AS IS" BASIS,
-/// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-/// See the License for the specific language governing permissions and
-/// limitations under the License.
-///
-/// Copyright holder is SereneDB GmbH, Berlin, Germany
-////////////////////////////////////////////////////////////////////////////////
-
-#include <type_traits>
-#include <utility>
-
-#include "iresearch/search/common/fill_posting_scored.hpp"
-#include "iresearch/search/fill/impl.hpp"
-#include "iresearch/search/fill/plan.hpp"
-#include "iresearch/search/fill/window_disjunction.hpp"
-
-namespace irs::fill {}  // namespace irs::fill
