+++ libs/iresearch/include/iresearch/search/probe/boolean_sparse.hpp
@@ -23,29 +23,49 @@
 #include <tuple>
 #include <type_traits>
 #include <utility>
+#include <vector>
 
 #include "basics/empty.hpp"
 #include "basics/shared.hpp"
+#include "iresearch/search/common/score/make_conjunction.hpp"
+#include "iresearch/search/common/score_args.hpp"
+#include "iresearch/search/common/score_policy.hpp"
+#include "iresearch/search/score_function.hpp"
 #include "iresearch/utils/type_limits.hpp"
 
 namespace irs::probe {
 
-template<typename Musts, typename Optional, typename Excludes>
+template<typename Optional>
+consteval bool OptionalDecides() {
+  if constexpr (requires { Optional::kDecides; }) {
+    return Optional::kDecides;
+  } else {
+    return true;
+  }
+}
+
+template<typename Musts, typename Optional, typename Excludes,
+         typename Score = utils::Empty>
 class BooleanSparse {
  public:
   static constexpr bool kMusts = !std::is_same_v<Musts, utils::Empty>;
   static constexpr bool kOptional = !std::is_same_v<Optional, utils::Empty>;
   static constexpr bool kExcludes = !std::is_same_v<Excludes, utils::Empty>;
+  static constexpr bool kScored = std::is_same_v<Score, search::Scored>;
+  static constexpr bool kInherited = std::is_same_v<Score, search::Inherited>;
   static_assert(kMusts || kOptional);
+  static_assert(!kInherited || (kMusts && !kOptional));
 
   template<typename MustsArgs, typename OptionalArgs, typename ExcludesArgs>
   BooleanSparse(std::piecewise_construct_t, MustsArgs&& musts,
-                OptionalArgs&& optional, ExcludesArgs&& excludes)
+                OptionalArgs&& optional, ExcludesArgs&& excludes,
+                Score score = {})
     : _musts{std::make_from_tuple<Musts>(std::forward<MustsArgs>(musts))},
       _optional{
         std::make_from_tuple<Optional>(std::forward<OptionalArgs>(optional))},
       _excludes{
-        std::make_from_tuple<Excludes>(std::forward<ExcludesArgs>(excludes))} {}
+        std::make_from_tuple<Excludes>(std::forward<ExcludesArgs>(excludes))},
+      _score{score} {}
 
   BooleanSparse(BooleanSparse&&) = delete;
   BooleanSparse& operator=(BooleanSparse&&) = delete;
@@ -69,10 +89,66 @@ class BooleanSparse {
     return target;
   }
 
+  IRS_FORCE_INLINE void FetchScoreArgs(uint32_t slot)
+    requires(kScored || kInherited)
+  {
+    if constexpr (kMusts) {
+      _musts.FetchScoreArgs(slot);
+    }
+    if constexpr (kOptional) {
+      _optional.FetchScoreArgs(slot);
+    }
+  }
+
+  ScoreFunction PrepareScore()
+    requires(kScored || kInherited)
+  {
+    if constexpr (kInherited) {
+      return _musts.PrepareScore();
+    } else if constexpr (!kMusts && OptionalDecides<Optional>()) {
+      return _optional.PrepareScore(_score.inner, _score.absorbed);
+    } else {
+      std::vector<ScoreFunction> scorers;
+      if constexpr (kMusts) {
+        _musts.CollectScorers(scorers);
+      }
+      if constexpr (kOptional) {
+        search::AppendScorer(scorers, OptionalScore());
+      }
+      return search::MakeConjunctionScore(_score.inner, std::move(scorers),
+                                          _score.absorbed);
+    }
+  }
+
+  void CollectScorers(std::vector<ScoreFunction>& out)
+    requires(kScored || kInherited)
+  {
+    if constexpr (kInherited) {
+      _musts.CollectScorers(out);
+    } else {
+      search::AppendScorer(out, PrepareScore());
+    }
+  }
+
  private:
+  ScoreFunction OptionalScore()
+    requires kScored
+  {
+    if constexpr (requires { _optional.PrepareScore(_score.inner); }) {
+      return _optional.PrepareScore(_score.inner);
+    } else if constexpr (requires {
+                           _optional.PrepareScore(_score.inner, score_t{0});
+                         }) {
+      return _optional.PrepareScore(_score.inner, score_t{0});
+    } else {
+      return _optional.PrepareScore();
+    }
+  }
+
   [[no_unique_address]] Musts _musts;
   [[no_unique_address]] Optional _optional;
   [[no_unique_address]] Excludes _excludes;
+  [[no_unique_address]] Score _score;
 };
 
 }  // namespace irs::probe
