+++ libs/iresearch/include/iresearch/search/probe/plan.hpp
@@ -28,12 +28,12 @@
 #include "iresearch/search/common/posting_probe.hpp"
 #include "iresearch/search/common/probe_leaves.hpp"
 #include "iresearch/search/common/resolve.hpp"
+#include "iresearch/search/common/score_policy.hpp"
 #include "iresearch/search/probe/boolean_sparse.hpp"
 #include "iresearch/search/probe/impl.hpp"
 #include "iresearch/search/probe/leaves.hpp"
 #include "iresearch/search/probe/make.hpp"
 #include "iresearch/search/probe/single_posting.hpp"
-#include "iresearch/search/probe/sparse_disjunction_scored.hpp"
 
 namespace irs::probe {
 
@@ -109,9 +109,12 @@ Node::ptr MakeSparseDisjunctionScored(
     [&]<typename Leaf>(size_t size, auto&& init) -> Node::ptr {
       return search::ResolveArity<search::kRunArity, search::kRunFloor>(
         size, [&]<size_t N> -> Node::ptr {
-          using Node = SparseDisjunctionScored<Leaf, N>;
+          using Node = BooleanSparse<utils::Empty, OrLeaves<Leaf, N, true>,
+                                     utils::Empty, search::Scored>;
           return memory::make_managed<Impl<Node>>(
-            size, std::forward<decltype(init)>(init), merge, absorbed);
+            std::piecewise_construct, std::forward_as_tuple(),
+            std::forward_as_tuple(size, std::forward<decltype(init)>(init)),
+            std::forward_as_tuple(), search::Scored{merge, absorbed});
         });
     },
     search::ProbeOrder::Densest);
